+++ packages/replay/src/constants.ts
@@ -25,3 +25,8 @@ export const MAX_SESSION_LIFE = 1_800_000; // 30 minutes
  */
 export const DEFAULT_SESSION_SAMPLE_RATE = 0.1;
 export const DEFAULT_ERROR_SAMPLE_RATE = 1.0;
+
+export const REPLAY_SDK_INFO = {
+  name: 'sentry.javascript.integration.replay',
+  version: __SENTRY_REPLAY_VERSION__,
+};