+++ packages/replay/src/types.ts
@@ -7,6 +7,8 @@ export type RecordedEvents = Uint8Array | string;
 
 export type AllPerformanceEntry = PerformancePaintTiming | PerformanceResourceTiming | PerformanceNavigationTiming;
 
+export type ReplayRecordingMode = 'session' | 'error';
+
 export interface SendReplay {
   events: RecordedEvents;
   replayId: string;
@@ -105,6 +107,18 @@ export interface ReplayPluginOptions extends SessionOptions {
    * Block all media (e.g. images, svg, video) in recordings.
    */
   blockAllMedia: boolean;
+
+  /**
+   * _experiments allows users to enable experimental or internal features.
+   * We don't consider such features as part of the public API and hence we don't guarantee semver for them.
+   * Experimental features can be added, changed or removed at any time.
+   *
+   * Default: undefined
+   */
+  _experiments?: Partial<{
+    captureExceptions: boolean;
+    traceInternals: boolean;
+  }>;
 }
 
 // These are optional for ReplayPluginOptions because the plugin sets default values
@@ -206,6 +220,7 @@ export interface ReplayContainer {
   eventBuffer: EventBuffer | null;
   performanceEvents: AllPerformanceEntry[];
   session: Session | undefined;
+  recordingMode: ReplayRecordingMode;
   isEnabled(): boolean;
   isPaused(): boolean;
   getContext(): InternalEventContext;
@@ -218,4 +233,5 @@ export interface ReplayContainer {
   flushImmediate(): void;
   triggerUserActivity(): void;
   addUpdate(cb: AddUpdateCallback): void;
+  getOptions(): ReplayPluginOptions;
 }