+++ packages/replay/test/mocks/mockSdk.ts
@@ -8,6 +8,7 @@ import type { ReplayConfiguration } from '../../src/types';
 export interface MockSdkParams {
   replayOptions?: ReplayConfiguration;
   sentryOptions?: BrowserOptions;
+  autoStart?: boolean;
 }
 
 class MockTransport implements Transport {
@@ -35,7 +36,7 @@ class MockTransport implements Transport {
   }
 }
 
-export async function mockSdk({ replayOptions, sentryOptions }: MockSdkParams = {}): Promise<{
+export async function mockSdk({ replayOptions, sentryOptions, autoStart = true }: MockSdkParams = {}): Promise<{
   replay: ReplayContainer;
   integration: ReplayIntegration;
 }> {
@@ -75,7 +76,10 @@ export async function mockSdk({ replayOptions, sentryOptions }: MockSdkParams = 
 
   // Instead of `setupOnce`, which is tricky to test, we call this manually here
   replayIntegration['_setup']();
-  replayIntegration.start();
+
+  if (autoStart) {
+    replayIntegration.start();
+  }
 
   const replay = replayIntegration['_replay']!;