+++ packages/replay/test/unit/index-sampling.test.ts
@@ -1,5 +1,3 @@
-jest.unmock('@sentry/browser');
-
 // mock functions need to be imported first
 import { mockRrweb, mockSdk } from './../index';
 import { useFakeTimers } from './../utils/use-fake-timers';