+++ packages/replay/test/unit/index.test.ts
@@ -1,7 +1,4 @@
-jest.mock('./../../src/util/isInternal', () => ({
-  isInternal: jest.fn(() => true),
-}));
-
+import { getCurrentHub } from '@sentry/core';
 import { EventType } from 'rrweb';
 
 import { MAX_SESSION_LIFE, REPLAY_SESSION_KEY, VISIBILITY_CHANGE_TIMEOUT, WINDOW } from '../../src/constants';
@@ -13,7 +10,7 @@ import { useFakeTimers } from '../utils/use-fake-timers';
 import { PerformanceEntryResource } from './../fixtures/performanceEntry/resource';
 import { BASE_TIMESTAMP, RecordMock } from './../index';
 import { resetSdkMock } from './../mocks/resetSdkMock';
-import { DomHandler, MockTransportSend } from './../types';
+import { DomHandler } from './../types';
 
 useFakeTimers();
 
@@ -96,9 +93,8 @@ describe('Replay with custom mock', () => {
 describe('Replay', () => {
   let replay: ReplayContainer;
   let mockRecord: RecordMock;
-  let mockTransportSend: MockTransportSend;
+  let mockTransportSend: jest.SpyInstance<any>;
   let domHandler: DomHandler;
-  let spyCaptureException: jest.MockedFunction<any>;
   const prevLocation = WINDOW.location;
 
   type MockSendReplayRequest = jest.MockedFunction<typeof replay.sendReplayRequest>;
@@ -110,11 +106,13 @@ describe('Replay', () => {
   });
 
   beforeEach(async () => {
-    ({ mockRecord, mockTransportSend, domHandler, replay, spyCaptureException } = await resetSdkMock({
+    ({ mockRecord, domHandler, replay } = await resetSdkMock({
       replayOptions: {
         stickySession: false,
       },
     }));
+
+    mockTransportSend = jest.spyOn(getCurrentHub().getClient()!.getTransport()!, 'send');
 
     jest.spyOn(replay, 'flush');
     jest.spyOn(replay, 'runFlush');
@@ -631,6 +629,9 @@ describe('Replay', () => {
     // Suppress console.errors
     const mockConsole = jest.spyOn(console, 'error').mockImplementation(jest.fn());
 
+    // Check errors
+    const spyHandleException = jest.spyOn(replay, 'handleException');
+
     expect(replay.session?.segmentId).toBe(0);
 
     // fail the first and second requests and pass the third one
@@ -662,11 +663,10 @@ describe('Replay', () => {
     expect(replay.sendReplayRequest).toHaveBeenCalledTimes(4);
     expect(replay.sendReplay).toHaveBeenCalledTimes(4);
 
-    expect(spyCaptureException).toHaveBeenCalledTimes(5);
     // Retries = 3 (total tries = 4 including initial attempt)
     // + last exception is max retries exceeded
-    expect(spyCaptureException).toHaveBeenCalledTimes(5);
-    expect(spyCaptureException).toHaveBeenLastCalledWith(new Error('Unable to send Replay - max retries exceeded'));
+    expect(spyHandleException).toHaveBeenCalledTimes(5);
+    expect(spyHandleException).toHaveBeenLastCalledWith(new Error('Unable to send Replay - max retries exceeded'));
 
     // No activity has occurred, session's last activity should remain the same
     expect(replay.session?.lastActivity).toBe(BASE_TIMESTAMP);